refactor(forge): split GitHub label methods into github_labels.go (#1025)
Pure same-package code motion: move ListLabels, listRepoLabels,
EnsureLabels, UpdateIssueLabels, and ListIssueLabelEvents verbatim out of
github.go into the new github_labels.go. sameNameSet stays in forgejo.go
(shared with the Forgejo driver); UpdateIssueLabels calls it cross-file.
No behaviour, signature, body, or doc-comment change.
